change send cmd to invoke
