+++ pkg/invoke/invoke.go
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package send
+package invoke
 
 import (
 	"bytes"
